moved appearance relation options to new section
